fixed bug with summary details Approved/Submitted/Draft not changing with compliance year

+++ src/EA.Weee.Integration.Tests/Builders/TransferEvidenceNoteDbSetup.cs
@@ -28,7 +28,7 @@
             user.UserId.ToString(),
             tonnages,
             categories,
-            (short)DateTime.Now.Year);
+            DateTime.Now.Year);
         }
 
         public TransferEvidenceNoteDbSetup WithStatus(NoteStatus statusToUpdate, string user)
